Update (Sheet-1)F - Digits Summation.java

+++ (Sheet-1)F - Digits Summation.java
@@ -3,8 +3,8 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int x = sc.nextInt();
-        int y = sc.nextInt();
+        long x = sc.nextLong();
+        long y = sc.nextLong();
         int sum = 0;
         int a, b;
         a = x % 10;
